% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidation.R
\name{crossvalidate_lssm}
\alias{crossvalidate_lssm}
\title{Obtain cross-validated performance for lssm models}
\usage{
crossvalidate_lssm(
  y,
  ts_frequency,
  initial_window = 150,
  horizon = 1,
  fixed_window = FALSE,
  crossval_frequency = 1,
  tune_grid,
  verbose = TRUE,
  parallel = FALSE
)
}
\description{
Obtain cross-validated performance for lssm models
}
