% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.R
\name{dbcmvnorm}
\alias{dbcmvnorm}
\title{Multivariate Box-Cox transformed normal probability density function}
\usage{
dbcmvnorm(x, mean = rep(0, p), sigma = diag(p), lambda = 0, offset = 0, log)
}
\arguments{
\item{x}{vector or matrix of quantiles.  If \code{x} is a matrix, each row is
taken to be a quantile.}

\item{mean}{mean vector, default is \code{rep(0, length = ncol(x))}}

\item{sigma}{covariance matrix, deault is \code{diag(ncol(x))}}

\item{lambda}{power for Box-Cox transformation}

\item{offset}{offset applied before Box-Cox transformation}

\item{log}{logical; if \code{TRUE}, densities \code{d} are given as \code{log(d)}}
}
\description{
Multivariate Box-Cox transformed normal probability density function
}
