% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.R
\name{dlogmvnorm}
\alias{dlogmvnorm}
\title{Multivariate log-normal probability density function}
\usage{
dlogmvnorm(x, mean = rep(0, p), sigma = diag(p), offset = 0, log)
}
\arguments{
\item{x}{vector or matrix of quantiles.  If \code{x} is a matrix, each row is
taken to be a quantile.}

\item{mean}{mean vector, default is \code{rep(0, length = ncol(x))}}

\item{sigma}{covariance matrix, deault is \code{diag(ncol(x))}}

\item{offset}{offset applied before log transformation}

\item{log}{logical; if \code{TRUE}, densities \code{d} are given as \code{log(d)}}
}
\description{
Multivariate log-normal probability density function
}
