% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{drop_leading_nas}
\alias{drop_leading_nas}
\title{Remove leading values that are infinite or missing}
\usage{
drop_leading_nas(y)
}
\arguments{
\item{y}{a univariate time series or numeric vector}
}
\value{
a vector of the same class as y
}
\description{
Remove leading values that are infinite or missing
}
