% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma.R
\name{fit_arma}
\alias{fit_arma}
\title{Fit a quantile baseline model to historical disease incidence}
\usage{
fit_arma(y, verbose = FALSE, ...)
}
\arguments{
\item{y}{numeric vector of response values}

\item{verbose}{logical: if TRUE, print output from estimation in Stan}

\item{...}{other arguments are ignored}
}
\value{
arma fit object
}
\description{
Fit a quantile baseline model to historical disease incidence
}
