% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.R
\name{fit_lssm_caret_wrapper}
\alias{fit_lssm_caret_wrapper}
\title{Wrapper around fit_lssm for use with caret::train}
\usage{
fit_lssm_caret_wrapper(x, y, param, ts_frequency = 1, verbose = FALSE, ...)
}
\arguments{
\item{x}{time series data to fit to}

\item{y}{ignored}

\item{param}{dataframe of one row of arguments to fit_lssm}

\item{...}{other arguments are ignored}
}
\value{
numeric vector of predictive medians with attributes:
\itemize{
\item family is a string with the parametric family, e.g. "norm"
\item other attributes are names of parameters for the parametric family
}
}
\description{
Wrapper around fit_lssm for use with caret::train
}
