% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarima.R
\name{fit_sarima_lssm}
\alias{fit_sarima_lssm}
\title{Fit a quantile baseline model to historical disease incidence}
\usage{
fit_sarima_lssm(
  y,
  include_intercept,
  p_ar,
  q_ma,
  P_ar,
  Q_ma,
  stationary = 1L,
  ts_frequency = 1L,
  init_par,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{y}{numeric vector of response values}

\item{include_intercept}{logical: if TRUE, include an intercept}

\item{p_ar}{integer order for non-seasonal autoregressive terms}

\item{q_ma}{integer order of non-seasonal moving average terms}

\item{P_ar}{integer order for seasonal autoregressive terms}

\item{Q_ma}{integer order of seasonal moving average terms}

\item{ts_frequency}{integer frequency of time series}

\item{init_par}{initial values for model parameters in estimation}

\item{verbose}{logical: if TRUE, print output from estimation in Stan}

\item{...}{other arguments are ignored}
}
\value{
sarima_lssm fit object
}
\description{
Fit a quantile baseline model to historical disease incidence
}
