% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.R
\name{int_seq}
\alias{int_seq}
\title{Create a sequence of integers}
\usage{
int_seq(from, to, max_len)
}
\arguments{
\item{from}{starting value for sequence}

\item{to}{ending value for sequence}

\item{max_len}{maximum length of sequence}
}
\value{
sequence of integers starting at from, ending at to, and of length
at most max_len, with a preference for smaller integers
}
\description{
Create a sequence of integers
}
