% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma.R
\name{is.arma}
\alias{is.arma}
\title{Check if object is of class arma}
\usage{
is.arma(object)
}
\arguments{
\item{object}{an object that may be a arma object}
}
\value{
boolean; whether object is inherits arma class
}
\description{
Check if object is of class arma
}
