% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarima.R
\name{is.sarima_lssm}
\alias{is.sarima_lssm}
\title{Check if object is of class sarima_lssm}
\usage{
is.sarima_lssm(object)
}
\arguments{
\item{object}{an object that may be a sarima_lssm object}
}
\value{
boolean; whether object is inherits sarima_lssm class
}
\description{
Check if object is of class sarima_lssm
}
