% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.R
\name{log_score_summary}
\alias{log_score_summary}
\title{Calculate the log score for forecasts in the format required by caret}
\usage{
log_score_summary(data, ...)
}
\arguments{
\item{data}{data frame with columns obs and pred.  pred must be in the form
returned by predict_lssm_caret_wrapper}

\item{...}{other arguments are ignored}
}
\description{
Calculate the log score for forecasts in the format required by caret
}
