% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lssm-package.R
\docType{package}
\name{lssm}
\alias{lssm}
\alias{lssm-package}
\title{lssm - Linear State Space Models}
\description{
Estimating and forecasting from linear state space models
}
\author{
Evan L. Ray \href{mailto:elray@umass.edu}{elray@umass.edu}

Serena Wang \href{mailto:wang225y@mtholyoke.edu}{wang225y@mtholyoke.edu}
}
