% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarima.R
\name{new_sarima_lssm}
\alias{new_sarima_lssm}
\title{Create a sarima_lssm object}
\usage{
new_sarima_lssm(
  param_estimates,
  include_intercept,
  p_ar,
  q_ma,
  P_ar,
  Q_ma,
  ts_frequency
)
}
\arguments{
\item{param_estimates}{model parameter estimates}

\item{include_intercept}{logical; if TRUE, include intercept}

\item{p_ar}{integer order for non-seasonal autoregressive terms}

\item{q_ma}{integer order of non-seasonal moving average terms}

\item{P_ar}{integer order for seasonal autoregressive terms}

\item{Q_ma}{integer order of seasonal moving average terms}

\item{ts_frequency}{integer frequency of time series}
}
\value{
sarima_lssm fit object
}
\description{
Create a sarima_lssm object
}
