% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma.R
\name{predict.arma}
\alias{predict.arma}
\title{Predict future disease incidence starting from the end of the training data.}
\usage{
\method{predict}{arma}(
  arma_fit,
  newdata,
  horizon,
  forecast_representation,
  quantile_levels = c(0.025, 0.25, 0.5, 0.75, 0.975),
  nsim = 1e+05
)
}
\arguments{
\item{arma_fit}{an arma fit object}

\item{horizon}{number of time steps forward to predict}

\item{forecast_representation}{string specifying approach to representing
forecast distributions. One of "named_dist", "sample", or "quantile";
see documentation of \code{represent_forecasts} for more detail.}

\item{quantile_levels}{numeric vector of quantile levels to use for
forecast_representation = "quantile"}

\item{nsim}{integer number of samples to use for
forecast_representation = "sample"}

\item{...}{other arguments are ignored}
}
\value{
representation of forecasts
}
\description{
Predict future disease incidence starting from the end of the training data.
}
