% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarima.R
\name{predict.sarima_lssm}
\alias{predict.sarima_lssm}
\title{Predict future disease incidence starting from the end of the training data.}
\usage{
\method{predict}{sarima_lssm}(
  sarima_lssm_fit,
  newdata,
  horizon,
  forecast_representation,
  joint = TRUE,
  quantile_levels = c(0.025, 0.25, 0.5, 0.75, 0.975),
  nsim = 1e+05
)
}
\arguments{
\item{sarima_lssm_fit}{an sarima_lssm fit object}

\item{horizon}{number of time steps forward to predict}

\item{forecast_representation}{string specifying approach to representing
forecast distributions. One of "named_dist", "sample", or "quantile";
see documentation of \code{represent_forecasts} for more detail.}

\item{joint}{logical; if TRUE, named distribution representation tracks the
joint distribution of y_t over all horizons.  if FALSE, we track only the
marginal distributions at each horizon.}

\item{quantile_levels}{numeric vector of quantile levels to use for
forecast_representation = "quantile"}

\item{nsim}{integer number of samples to use for
forecast_representation = "sample"}

\item{...}{other arguments are ignored}
}
\value{
representation of forecasts
}
\description{
Predict future disease incidence starting from the end of the training data.
}
