% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represent.R
\name{represent_forecasts}
\alias{represent_forecasts}
\title{Translate representation of forecast distributions to a standardized form}
\usage{
represent_forecasts(
  named_dist_forecast,
  forecast_representation = c("named_dist", "sample", "quantile"),
  quantile_levels = c(0.025, 0.25, 0.5, 0.75, 0.975),
  nsim = 1e+05
)
}
\arguments{
\item{named_dist_forecast}{data frame specifying the forecast, with at
minimum a "family" column. Any columns occurring before the "family"
are assumed to be identifiers of the forecast, such as a forecast horizon or
location; these columns are retained in the output. Any columns after the
"family" are assumed to be parameters of the forecast distribution, and are
used as part of calls to the q or d function for the family.}

\item{forecast_representation}{string specifying approach to representing
forecast distributions, see documentation of represent_forecasts}

\item{quantile_levels}{numeric vector of quantile levels to use for
forecast_representation = "quantile"}

\item{nsim}{integer number of samples to use for
forecast_representation = "sample"}
}
\value{
a representation of the forecast; the structure depends on
\code{forecast_representation}:
\itemize{
\item "named_dist": the \code{named_dist_forecast} argument is returned
\item "sample": a matrix is returned with number of rows equal to the number of
rows of \code{named_dist_forecast} and number of columns equal to \code{nsim}.
Each row contains \code{nsims} samples from the predictive distribution in the
corresponding row of \code{named_dist_forecast}.
\item "quantile": a matrix is returned with number of rows equal to the number
of rows of \code{named_dist_forecast} and number of columns equal to the length
of \code{quantile_levels}. Each row contains quantiles of the predictive
distribution in the corresponding row of \code{named_dist_forecast}.
}
}
\description{
Translate representation of forecast distributions to a standardized form
}
