% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma.R
\name{new_arma}
\alias{new_arma}
\title{Create a arma object}
\usage{
new_arma(param_estimates)
}
\arguments{
\item{param_estimates}{model parameter estimates}

\item{p}{autoregressive order}
}
\value{
arma fit object
}
\description{
Create a arma object
}
