% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represent.r
\name{represent_forecasts}
\alias{represent_forecasts}
\title{Standardize representation of forecast distributions}
\usage{
represent_forecasts(
  named_dist_forecast,
  forecast_representation = c("named_dist", "sample", "quantile"),
  p = c(0.025, 0.25, 0.5, 0.75, 0.975),
  nsim = 1e+05
)
}
\arguments{
\item{named_dist_forecast}{data frame specifying the forecast, with at
minimum a "family" column. Any columns occurring before the "family"
are assumed to be identifiers of the forecast, such as a forecast horizon or
location; these columns are retained in the output. Any columns after the
"family" are assumed to be parameters of the forecast distribution, and are
used as part of calls to the q or d function for the family.}

\item{forecast_representation}{string specifying approach to representing
forecast distributions, see documentation of represent_forecasts}

\item{p}{numeric vector of quantile levels to use for
forecast_representation = "quantile"}

\item{nsim}{integer number of samples to use for
forecast_representation = "sample"}
}
\value{
a (potentially nested) tibble representing the forecast.
}
\description{
Standardize representation of forecast distributions
}
